requirements.txt
# This is a pip requirements file. Install it with `pip -r requirements.txt'

#simplejson # for python 2.5

# Django dependencies
#MySQL-python # For MySQL support, requires libmysqlclient-dev package to be installed
PIL

Django
-e git://github.com/rnk/django-media-bundler.git#egg=django-media-bundler

# video-enclave
pyOpenSSL # for eventlet
eventlet
mechanize
BeautifulSoup

# audio-enclave
Pyro
mutagen
python-daemon
